--- dd-java-agent/instrumentation/graal/graal-native-image-20.0/src/main/java/datadog/trace/instrumentation/graal/nativeimage/AbstractNativeImageInstrumentation.java
@@ -1,5 +1,0 @@
-package datadog.trace.instrumentation.graal.nativeimage;
-
-import datadog.trace.agent.tooling.Instrumenter;
-
-public abstract class AbstractNativeImageInstrumentation implements Instrumenter.HasMethodAdvice {}
