+++ dd-java-agent/instrumentation/graal/graal-native-image-20.0/src/main/java/datadog/trace/instrumentation/graal/nativeimage/AbstractNativeImageModule.java
@@ -1,0 +1,16 @@
+package datadog.trace.instrumentation.graal.nativeimage;
+
+import datadog.trace.agent.tooling.InstrumenterModule;
+import datadog.trace.api.Platform;
+import java.util.Set;
+
+public abstract class AbstractNativeImageModule extends InstrumenterModule {
+  public AbstractNativeImageModule() {
+    super("native-image");
+  }
+
+  @Override
+  public boolean isApplicable(Set<TargetSystem> enabledSystems) {
+    return Platform.isNativeImageBuilder();
+  }
+}
