+++ dd-java-agent/instrumentation/graal/graal-native-image-20.0/src/main/java/datadog/trace/instrumentation/graal/nativeimage/AnnotationSubstitutionProcessorInstrumentation.java
@@ -9,7 +9,7 @@ import java.util.List;
 import net.bytebuddy.asm.Advice;
 
 public final class AnnotationSubstitutionProcessorInstrumentation
-    extends AbstractNativeImageInstrumentation implements Instrumenter.ForSingleType {
+    implements Instrumenter.ForSingleType, Instrumenter.HasMethodAdvice {
 
   @Override
   public String instrumentedType() {
